+++ app/page.tsx
@@ -1,24 +1,15 @@
 "use client"
 
-import { useContext, useEffect, useState } from "react"
-import Link from "next/link"
-import { ArrowBigUp, ChevronDown, ChevronUp, RefreshCw } from "lucide-react"
-import { formatUnits } from "viem"
+import { useEffect, useState } from "react"
+import { useAppContext } from "@/context/AppContext"
+import { RefreshCw } from "lucide-react"
 import { useAccount } from "wagmi"
 
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert"
 import { Button } from "@/components/ui/button"
 import { Skeleton } from "@/components/ui/skeleton"
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/ui/table"
+import BalancesTable from "@/components/BalancesTable"
 import Swap from "@/components/swap"
-import { AppContext } from "@/app/index"
 
 const LoadingSkeleton = () => {
   return (
@@ -28,83 +19,6 @@ const LoadingSkeleton = () => {
         .map((_, index) => (
           <Skeleton key={index} className="h-10 w-full" />
         ))}
-    </div>
-  )
-}
-
-const BalancesTable = ({
-  balances,
-  showAll,
-  toggleShowAll,
-  stakingAmountTotal,
-}: any) => {
-  return (
-    <div>
-      <Table>
-        <TableHeader>
-          <TableRow className="border-none hover:bg-transparent">
-            <TableHead className="pl-4">Asset</TableHead>
-            <TableHead>Type</TableHead>
-            <TableHead className="text-right">Price</TableHead>
-            <TableHead className="text-right">Balance</TableHead>
-          </TableRow>
-        </TableHeader>
-        <TableBody>
-          {balances
-            .slice(0, showAll ? balances.length : 5)
-            .map((b: any, index: any) => (
-              <TableRow key={index} className="border-none">
-                <TableCell className="pl-4 rounded-bl-xl rounded-tl-xl">
-                  <div>{b.ticker}</div>
-                  <div className="text-xs text-gray-400">{b.chain_name}</div>
-                </TableCell>
-                <TableCell>{b.coin_type}</TableCell>
-                <TableCell className="text-right">
-                  {b.price?.toFixed(2)}
-                </TableCell>
-                <TableCell className="rounded-br-xl rounded-tr-xl text-right">
-                  {parseFloat(b.balance).toFixed(2) || "N/A"}
-                  {b.ticker === "ZETA" && b.coin_type === "Gas" && (
-                    <div>
-                      <Button
-                        size="sm"
-                        variant="link"
-                        className="my-1 p-0 text-xs h-5"
-                        asChild
-                      >
-                        <Link href="/staking">
-                          <ArrowBigUp className="h-4 w-4 mr-0.5" />
-                          {stakingAmountTotal > 0 ? (
-                            <span>
-                              Staked:&nbsp;
-                              {parseFloat(
-                                formatUnits(stakingAmountTotal, 18)
-                              ).toFixed(0)}
-                            </span>
-                          ) : (
-                            <span>Stake ZETA</span>
-                          )}
-                        </Link>
-                      </Button>
-                    </div>
-                  )}
-                </TableCell>
-              </TableRow>
-            ))}
-        </TableBody>
-      </Table>
-      {balances?.length > 5 && (
-        <div className="my-4 flex justify-center">
-          <Button variant="link" onClick={toggleShowAll}>
-            {showAll ? "Collapse" : "Show all assets"}
-            {showAll ? (
-              <ChevronUp className="ml-2 h-4 w-4 shrink-0 opacity-75" />
-            ) : (
-              <ChevronDown className="ml-2 h-4 w-4 shrink-0 opacity-75" />
-            )}
-          </Button>
-        </div>
-      )}
     </div>
   )
 }
@@ -128,7 +42,7 @@ export default function IndexPage() {
     fetchBalances,
     prices,
     stakingDelegations,
-  } = useContext(AppContext)
+  } = useAppContext()
   const [sortedBalances, setSortedBalances] = useState([])
   const [showAll, setShowAll] = useState(false)
 
@@ -144,12 +58,10 @@ export default function IndexPage() {
 
   const balancesPrices = sortedBalances.map((balance: any) => {
     const normalizeSymbol = (symbol: string) => symbol.replace(/^[tg]/, "")
-
     const normalizedSymbol = normalizeSymbol(balance.symbol)
     const priceObj = prices.find(
       (price: any) => normalizeSymbol(price.symbol) === normalizedSymbol
     )
-
     return {
       ...balance,
       price: priceObj ? priceObj.price : null,
@@ -167,20 +79,18 @@ export default function IndexPage() {
         // Prioritize ZETA
         if (a.ticker === "ZETA" && a.coin_type === "Gas") return -1
         if (b.ticker === "ZETA" && b.coin_type === "Gas") return 1
-
         if (a.coin_type === "Gas" && b.coin_type !== "Gas") return -1
         if (a.coin_type !== "Gas" && b.coin_type === "Gas") return 1
         return a.chain_name < b.chain_name ? -1 : 1
       })
-      .filter((b: any) => {
-        return b.balance > 0
-      })
+      .filter((b: any) => b.balance > 0)
     setSortedBalances(balance)
   }, [balances])
 
-  const balancesTotal = balancesPrices.reduce((a: any, c: any) => {
-    return a + parseFloat(c.balance)
-  }, 0)
+  const balancesTotal = balancesPrices.reduce(
+    (a: any, c: any) => a + parseFloat(c.balance),
+    0
+  )
 
   const formatBalanceTotal = (b: string) => {
     if (parseFloat(b) > 1000) {
